no problems

+++ src/main/java/cap/s42academy/model/Income.java
@@ -20,6 +20,6 @@ public class Income {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long incomeId;
     private LocalDateTime transactionTime;
-    private BigDecimal incomeAmount;
+    private Integer incomeAmount;
 
 }
